@@ -70,12 +70,12 @@
     "react-csv": "^2.2.2",
     "react-dom": "18.2.0",
     "react-helmet-async": "^2.0.4",
-    "react-intl": "^5.10.15",
+    "react-intl": "^6.4.4",
     "react-is": "18.2.0",
     "react-moment": "^1.1.1",
     "react-number-format": "^4.9.1",
     "react-router-dom": "^6.26.0",
-    "react-scroll": "^1.8.9",
+    "react-scroll": "^1.9.0",
     "react-slick": "^0.29.0",
     "react-swipeable": "^7.0.0",
     "react-use-intercom": "^5.4.0",
@@ -90,7 +90,7 @@
     "slick-carousel": "^1.8.1",
     "styled-components": "5.3.6",
     "tslib": "^2.3.0",
-    "use-query-params": "^2.1.1",
+    "use-query-params": "^2.2.1",
     "yarn": "^1.22.19"
   },
   "devDependencies": {
@@ -144,8 +144,8 @@
     "@types/react": "18.2.0",
     "@types/react-dom": "18.2.0",
     "@types/react-is": "17.0.3",
-    "@types/react-scroll": "^1.8.6",
-    "@types/react-transition-group": "^4.4.5",
+    "@types/react-scroll": "^1.8.10",
+    "@types/react-transition-group": "^4.4.10",
     "@types/sanitize-html": "^2.6.2",
     "@types/set-interval-async": "^1.0.0",
     "@types/styled-components": "5.1.26",
@@ -202,12 +202,15 @@
     "ts-jest": "28.0.5",
     "ts-node": "10.9.1",
     "typechain": "^8.3.2",
-    "typescript": "4.8.4",
+    "typescript": "^5.0.4",
     "url": "^0.11.0",
     "url-loader": "^4.1.1",
     "webpack": "^5.75.0",
     "webpack-merge": "^5.8.0",
     "wrangler": "^3.61.0"
+  },
+  "resolutions": {
+    "@types/react": "18.2.0"
   },
   "workspaces": [
     "packages/**",
