@@ -1,5 +1,5 @@
 import { useCallback, useEffect } from 'react';
-import { trackEvent, RouteState } from '@notional-finance/helpers';
+import { trackEvent } from '@notional-finance/helpers';
 import { TRACKING_EVENTS } from '@notional-finance/util';
 import { Divider, styled, useTheme, Box } from '@mui/material';
 import { useLocation } from 'react-router';
@@ -42,7 +42,7 @@ export const TransactionConfirmation = ({
 }: TransactionConfirmationProps) => {
   const theme = useTheme();
   const { state, updateState } = context;
-  const location = useLocation<RouteState>();
+  const location = useLocation();
   const {
     populatedTransaction,
     transactionError,
@@ -87,7 +87,7 @@ export const TransactionConfirmation = ({
               'By submitting a trade on our platform you agree to our <a>terms of service.</a>'
             }
             values={{
-              a: (msg: string) => (
+              a: (msg: React.ReactNode) => (
                 <ExternalLink
                   href="/terms"
                   style={{ color: theme.palette.primary.light }}
